fix: moved App.tsx contents
