+++ frontend/package.json
@@ -58,7 +58,5 @@
     "vite-tsconfig-paths": "^3.5.1",
     "vitest": "^0.25.2",
     "spotify-web-api-js": "1.5.2"
-
-    
   }
 }
