+++ frontend/src/pages/Index.tsx
@@ -2,6 +2,11 @@ import { QueueListIcon } from '@heroicons/react/24/outline'
 import { Link } from 'react-router-dom'
 import { Pie, PieChart } from 'recharts'
 
+import UserLogin from 'components/UserLogin'
+import { getTokenFromUrl, getPlayLists } from './lib/spotify'
+import SpotifyWebApi from 'spotify-web-api-js'
+import { useEffect, useState } from 'react'
+
 const Index: React.FC = () => {
   const data01 = [
     {
@@ -29,6 +34,43 @@ const Index: React.FC = () => {
       value: 189
     }
   ]
+
+  const spotify = new SpotifyWebApi()
+  const [spotifyToken, setSpotifyToken] = useState('')
+
+  useEffect(() => {
+    const function_a = async () => {
+      console.log('This is what we derived from the URL: ', getTokenFromUrl())
+      const _spotifyToken = getTokenFromUrl().access_token
+      window.location.hash = ''
+      console.log('This is our spotify token ', _spotifyToken)
+      if (_spotifyToken) {
+        setSpotifyToken(_spotifyToken)
+        spotify.setAccessToken(_spotifyToken)
+        const playlists = await spotify
+          .getUserPlaylists() // note that we don't pass a user id
+          .then(
+            function (data) {
+              const string = JSON.parse(JSON.stringify(data.items))
+              return string
+            },
+            function (err) {
+              console.error(err)
+            }
+          )
+        const user_Info = await spotify.getMe()
+        const user_Id = user_Info.id
+        console.log('The user Id we get is', user_Id)
+        console.log('playlists, ', playlists)
+        const myAsynFunction = async (playlists): Promise<T> => {
+          const value = await playlists
+          return value
+        }
+      }
+    }
+
+    function_a()
+  })
 
   return (
     <>
@@ -61,6 +103,7 @@ const Index: React.FC = () => {
             />
           </PieChart>
         </div>
+        <UserLogin />
       </div>
     </>
   )
